made changes in add state, zones and rpc, added token for api requests. Made changes in axios.js file.
