+++ webapp/src/constants/RouteConstants.js
@@ -7,7 +7,7 @@ export const DASHBOARD_URL = "/";
 export const ADD_USER = "/add-user";
 export const VIEW_STATES = "/states";
 export const VIEW_ZONES = "/view-zones";
-export const VIEW_RPC = "/rpc";
+export const VIEW_RPC = "/view-rpc";
 export const ADD_STATES = "/add-states";
 export const ADD_ZONES = "/add-zones";
 export const ADD_RPC = "/add-rpc";
