+++ webapp/src/constants/StrapiApiConstants.js
@@ -1,5 +1,5 @@
 export const STRAPI_LOGIN_PATH = "auth/local";
-export const STRAPI_DB_URL = "http://192.168.2.87:1337/";
+export const STRAPI_DB_URL = "http://localhost:1337/";
 export const STRAPI_ZONES = "zones";
 export const STRAPI_STATES = "states";
 export const STRAPI_RPCS = "rpcs";
