--- webapp/src/containers/RPC/ViewRpc.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-export default class ViewRpc extends React.Component {
-  render() {
-    return <div></div>;
-  }
-}
